+++ app/build.gradle.kts
@@ -42,7 +42,9 @@ android {
 
 dependencies {
     implementation(project(Dep.Modules.CORE_DESIGN_SYSTEM))
-    implementation(project(Dep.Modules.FEATURE))
+    implementation(project(Dep.Modules.FEATURE_HOME))
+    implementation(project(Dep.Modules.FEATURE_RECORD))
+    implementation(project(Dep.Modules.FEATURE_INTRODUCE))
     implementation(project(Dep.Modules.CORE_NAVIGATOR))
 
     implementation(Dep.Hilt.HILT)
