+++ buildSrc/src/main/kotlin/Dep.kt
@@ -9,7 +9,10 @@ object Dep {
         const val CORE_NETWORK = ":core-network"
         const val CORE_DATA = ":core-data"
 
-        const val FEATURE = ":feature"
+        const val FEATURE_HOME = ":feature:home"
+        const val FEATURE_RECORD = ":feature:record"
+        const val FEATURE_INTRODUCE = ":feature:introduce"
+        const val FEATURE_ONBOARDING = ":feature:onboarding"
     }
 
     object Gradle {
